Polish Global Exception handler and add exception handling for delete method in Service layer

+++ QamarRaza_java_training/session5/taskmanager/src/main/java/com/springadvanced/taskmanager/service/TodoService.java
@@ -67,7 +67,14 @@ public class TodoService
     //Delete a task via DTO
     public void deleteTodo(Long id)
     {
-        todoRepository.deleteById(id);
+        if(!todoRepository.existsById(id))
+        {
+            throw new TodoNotFoundException(id);
+        }
+        else
+        {
+            todoRepository.deleteById(id);
+        }
     }
 
 
@@ -114,7 +121,7 @@ public class TodoService
 
         else
         {
-            throw new TodoNotFoundException();
+            throw new TodoNotFoundException(id);
         }
     }
